[BE-031] Implement JPA Book and Repository Adapter (#82)
* Feat (BE-031): add BookEntity and mapper

* Feat (BE-031): add BookPersistenceMapperTest

* Feat (BE-031): add JpaBookRepository

* Feat (BE-031): add JpaBookRepositoryIT

+++ src/main/kotlin/org/cescfe/bookpublishing/book/domain/port/BookRepositoryView.kt
@@ -14,7 +14,7 @@ interface BookRepositoryView {
         limit: Int,
     ): List<BookSummary>
 
-    fun countAll(): Int
+    fun countAll(): Long
 
     fun save(book: Book): Book
